refactor(harvest): 4 client classes 抽离为 harvest_clients/ 子包 (#86)
harvest.py 3974→2425 行，4 条协议路径各自独立模块化：
- base.py: HTTP helpers + SSE streaming + retry infra
- gateway.py: GatewayClient (OpenAI chat/completions)
- anthropic.py: AnthropicGatewayClient + cache control
- responses.py: ResponsesGatewayClient
- gemini.py: GeminiNativeGatewayClient
- factory.py: make_client_factory

零行为变更，474 测试全绿。
