+++ plugins/deep-research/scripts/harvest.py
@@ -416,311 +416,60 @@ def build_rejected_claims(invalid):
 
 
 # ---------------------------------------------------------------------------
-# HTTP helpers
+# HTTP/SSE/client infrastructure -- extracted to harvest_clients/ (pure
+# refactor, zero behavior change). Re-imported here so every existing
+# harvest.<name> call site and test reference keeps working unchanged.
 # ---------------------------------------------------------------------------
-
-def _maybe_decompress(raw, resp):
-    encoding = (resp.headers.get("Content-Encoding") or "").lower()
-    if encoding == "gzip":
-        try:
-            return gzip.decompress(raw)
-        except OSError:
-            return raw
-    return raw
-
-
-def _http_json_post(url, payload, headers, timeout):
-    data = json.dumps(payload).encode("utf-8")
-    req = urllib.request.Request(url, data=data, headers=headers, method="POST")
-    with urllib.request.urlopen(req, timeout=timeout) as resp:
-        raw = resp.read()
-        raw = _maybe_decompress(raw, resp)
-        return json.loads(raw.decode("utf-8"))
-
-
-# Retry schedule for the gateway chat/completions call only (shared by the
-# panel loop and the judge) -- other backends keep their existing
-# degrade-to-next-backend behavior and are not retried here.
-_GATEWAY_RETRY_BACKOFFS_TRANSIENT = (5, 15)  # 429/408/5xx: up to 2 retries
-_GATEWAY_RETRY_BACKOFFS_OTHER_4XX = (5,)     # other 4xx (401/403/404...): 1 retry
-_COMPLETION_RETRY_BACKOFFS = (2, 5, 15, 30)  # 4 retries, 52s total backoff
-
-
-def _http_json_post_with_retry(url, payload, headers, timeout, transient_backoffs=None):
-    """Bounded retry around _http_json_post for transient gateway failures.
-    Classification is fixed on the first failure's status code and the
-    matching backoff schedule is followed to exhaustion -- no lock is held
-    across this function or its sleeps, so worker/judge threads never block
-    each other while backing off.
-
-    Two failure shapes share one retry loop: an HTTP response with a status
-    code (HTTPError -- classified transient/4xx by status), and a network-
-    layer failure with no status code at all (URLError/socket.timeout --
-    always treated as transient, same as a 5xx). HTTPError must be caught
-    before URLError since urllib raises HTTPError as a URLError subclass;
-    catching the parent first would swallow the status-code-bearing case."""
-    backoffs = None
-    attempt = 0
-    while True:
-        attempt += 1
-        try:
-            return _http_json_post(url, payload, headers, timeout)
-        except urllib.error.HTTPError as e:
-            status = e.code
-            if backoffs is None:
-                transient = status in (429, 408) or 500 <= status < 600
-                backoffs = (transient_backoffs if transient_backoffs is not None else _GATEWAY_RETRY_BACKOFFS_TRANSIENT) if transient else _GATEWAY_RETRY_BACKOFFS_OTHER_4XX
-            retries_done = attempt - 1
-            if retries_done >= len(backoffs):
-                raise RuntimeError(f"HTTP {status} after {attempt} attempts") from e
-            time.sleep(backoffs[retries_done])
-        except (urllib.error.URLError, socket.timeout, TimeoutError, json.JSONDecodeError) as e:
-            # No status code (connection refused/reset, DNS failure, read
-            # timeout, ...) -- always transient, same schedule as 429/5xx.
-            if backoffs is None:
-                backoffs = transient_backoffs if transient_backoffs is not None else _GATEWAY_RETRY_BACKOFFS_TRANSIENT
-            retries_done = attempt - 1
-            if retries_done >= len(backoffs):
-                raise RuntimeError(f"network error after {attempt} attempts: {e}") from e
-            time.sleep(backoffs[retries_done])
-
-
-# ---------------------------------------------------------------------------
-# SSE streaming infrastructure (fail-fast watchdog for completion calls)
-#
-# Pure protocol layer: _http_stream_post knows nothing about any provider's
-# event shapes, only the generic SSE wire format (event:/data:/comment lines
-# terminated by a blank line). Each of the four completion clients' complete()
-# consumes this generator and layers its own provider-specific frame parsing
-# and logical-end detection on top -- see AnthropicGatewayClient.complete /
-# ResponsesGatewayClient.complete / GeminiNativeGatewayClient.complete /
-# GatewayClient.complete.
-# ---------------------------------------------------------------------------
-
-class StreamIdleTimeout(RuntimeError):
-    """Inter-token or TTFT (time-to-first-token) timeout -- transient,
-    retryable under the same backoff schedule as a network-layer failure."""
-
-
-class StreamConnectionLost(RuntimeError):
-    """Mid-stream connection drop (socket reset, truncated read, TLS error)
-    -- transient, retryable."""
-
-
-class StreamNotSupportedError(RuntimeError):
-    """Server returned HTTP 200 but the body is not an SSE stream (e.g. a
-    plain application/json error payload) -- permanent, callers must NOT
-    retry this as a transient stream failure. The escape-hatch config flags
-    (limits.anthropic_stream etc.) exist precisely so an operator can flip
-    back to the non-streaming path if a gateway turns out not to support
-    streaming at all."""
-
-
-def _get_raw_socket(resp):
-    """Best-effort extraction of the underlying socket from an
-    http.client.HTTPResponse so a caller can settimeout() it once streaming
-    is under way (tightening from a generous TTFT timeout to a tighter
-    inter-token timeout). Tries the shapes seen across CPython's http.client/
-    ssl-wrapped-socket implementations; returns None (never raises) if none
-    match, so a caller that can't get a socket degrades to "no watchdog
-    tightening" rather than crashing."""
-    for attr_chain in [('fp', 'raw', '_sock'), ('fp', '_sock'), ('sock',)]:
-        obj = resp
-        for attr in attr_chain:
-            obj = getattr(obj, attr, None)
-            if obj is None:
-                break
-        else:
-            if hasattr(obj, 'settimeout'):
-                return obj
-    return None
-
-
-# Exceptions that can surface from resp.readline()/resp.read() once the
-# connection is open -- collapsed into StreamIdleTimeout at the readline
-# call site (the caller distinguishes idle-timeout from connection-lost by
-# *when* in the frame lifecycle the exception occurred, not by exception
-# type -- both socket.timeout from an inter-token settimeout() and a genuine
-# mid-read connection reset raise through this same set on CPython).
-_STREAM_READ_EXCEPTIONS = (
-    http.client.IncompleteRead, urllib.error.URLError, ConnectionResetError,
-    TimeoutError, socket.timeout, ssl.SSLError,
+from harvest_clients import (
+    GatewayClient,
+    AnthropicGatewayClient,
+    ResponsesGatewayClient,
+    GeminiNativeGatewayClient,
+    make_client_factory,
 )
-
-
-def _http_stream_post(url, payload, headers, initial_timeout):
-    """POST `payload` and yield parsed Server-Sent-Events frames.
-
-    Usage::
-
-        stream = _http_stream_post(url, payload, headers, timeout)
-        resp = next(stream)            # underlying HTTPResponse
-        sock = _get_raw_socket(resp)   # may be None
-        for event_type, data_str in stream:
-            ...
-
-    Yields:
-        - First value: the raw HTTPResponse object, so the caller can obtain
-          the underlying socket via _get_raw_socket() before consuming any
-          frames.
-        - Every value after that: an ``(event_type, data_str)`` tuple, one
-          per complete SSE frame (a run of ``event:``/``data:`` lines
-          terminated by a blank line; multiple ``data:`` lines in one frame
-          are newline-joined per the SSE spec). A bare comment line (leading
-          ``:``) yields ``(None, None)`` immediately as a heartbeat -- it
-          does not participate in frame buffering.
-
-    Raises:
-        urllib.error.HTTPError: the initial connection attempt got a non-2xx
-            response. Re-raised as-is (not wrapped) so a caller's retry loop
-            can classify it exactly like `_http_json_post_with_retry` /
-            `_anthropic_post_with_retry` already do (429/408/5xx transient,
-            other 4xx not) -- the response body is stashed on
-            `e.stream_error_body` (best-effort, mirroring
-            `_read_http_error_body`'s pattern) since `e.read()` can only be
-            called once.
-        StreamNotSupportedError: the connection succeeded (HTTP 200) but the
-            response Content-Type is not text/event-stream (e.g. the gateway
-            fell back to a plain JSON error body) -- permanent, do not retry
-            as a stream failure.
-        StreamConnectionLost: reading the non-SSE response body itself failed
-            at the socket level, or a mid-stream error surfaces at the
-            protocol layer.
-        StreamIdleTimeout: a readline() call raised a socket-level exception
-            while consuming the event stream (covers both "no bytes for N
-            seconds" once the caller has tightened the socket timeout, and a
-            genuine connection drop -- the caller's retry loop treats both as
-            transient either way).
-
-    The whole body runs inside a ``with`` block around the connection-opening
-    call below, so a consumer that stops iterating early (GeneratorExit, e.g.
-    from ``contextlib.closing``) still closes the underlying connection via
-    the context manager's __exit__.
-    """
-    data = json.dumps(payload).encode("utf-8")
-    req = urllib.request.Request(url, data=data, headers=headers, method="POST")
-    try:
-        with urllib.request.urlopen(req, timeout=initial_timeout) as resp:
-            content_type = (resp.headers.get("Content-Type") or "").lower()
-            if "text/event-stream" not in content_type:
-                try:
-                    body = resp.read(500)
-                except _STREAM_READ_EXCEPTIONS as e:
-                    raise StreamConnectionLost(
-                        f"failed reading non-stream response body "
-                        f"(Content-Type={content_type!r}): {e}"
-                    ) from e
-                body_text = body.decode("utf-8", errors="replace")
-                raise StreamNotSupportedError(
-                    f"expected text/event-stream, got Content-Type={content_type!r}: {body_text}"
-                )
-
-            yield resp
-
-            event_type = None
-            data_lines = []
-            while True:
-                try:
-                    raw_line = resp.readline()
-                except _STREAM_READ_EXCEPTIONS as e:
-                    raise StreamIdleTimeout(f"stream read failed: {e}") from e
-
-                if raw_line == b"":
-                    # EOF -- flush a buffered partial frame (if any), then stop.
-                    if data_lines or event_type is not None:
-                        yield (event_type, "\n".join(data_lines))
-                    return
-
-                line = raw_line.decode("utf-8", errors="replace").rstrip("\r\n")
-
-                if line == "":
-                    # Blank line terminates the current frame (SSE spec) --
-                    # yield unconditionally, even if the frame is empty (a
-                    # stray keepalive blank line); callers already skip
-                    # empty/whitespace-only data_str per the shared
-                    # consumption contract.
-                    yield (event_type, "\n".join(data_lines))
-                    event_type = None
-                    data_lines = []
-                    continue
-
-                if line.startswith(":"):
-                    # Comment/heartbeat line -- does not belong to any frame.
-                    yield (None, None)
-                    continue
-
-                if line.startswith("event:"):
-                    event_type = line[len("event:"):].strip()
-                    continue
-
-                if line.startswith("data:"):
-                    # SSE spec: strip at most one leading space after "data:".
-                    field_value = line[len("data:"):]
-                    if field_value.startswith(" "):
-                        field_value = field_value[1:]
-                    data_lines.append(field_value)
-                    continue
-                # Other SSE fields (id:, retry:) carry no meaning for any of
-                # the four providers this module speaks to -- ignored.
-    except urllib.error.HTTPError as e:
-        # Re-raise (not wrapped in RuntimeError) so the caller's retry loop
-        # can classify by e.code exactly like the non-streaming paths do.
-        # Stash the body since e.read() is single-shot and the caller may
-        # want it for a final error message after retries are exhausted.
-        e.stream_error_body = _read_http_error_body(e)
-        raise
-
-
-# Exceptions a streaming completion attempt can raise that the retry loop
-# below treats as transient (same set specified for all four streaming
-# clients): idle/connection failures at the SSE layer, raw socket-level
-# timeouts (a caller's own settimeout() can raise these directly, not just
-# via _STREAM_READ_EXCEPTIONS inside _http_stream_post), and a malformed
-# frame body (JSONDecodeError) -- one garbled chunk is worth a resend, not a
-# hard failure. urllib.error.HTTPError is handled separately (by status
-# code) rather than blanket-listed here.
-_STREAM_RETRYABLE_EXCEPTIONS = (
-    StreamIdleTimeout, StreamConnectionLost, socket.timeout, TimeoutError,
-    urllib.error.URLError, json.JSONDecodeError, ssl.SSLError,
+from harvest_clients.base import (
+    _read_http_error_body,
+    _maybe_decompress,
+    _http_json_post,
+    _GATEWAY_RETRY_BACKOFFS_TRANSIENT,
+    _GATEWAY_RETRY_BACKOFFS_OTHER_4XX,
+    _COMPLETION_RETRY_BACKOFFS,
+    _http_json_post_with_retry,
+    StreamIdleTimeout,
+    StreamConnectionLost,
+    StreamNotSupportedError,
+    _get_raw_socket,
+    _STREAM_READ_EXCEPTIONS,
+    _http_stream_post,
+    _STREAM_RETRYABLE_EXCEPTIONS,
+    _run_stream_attempt_with_retry,
 )
-
-
-def _run_stream_attempt_with_retry(attempt_fn, backoffs=_COMPLETION_RETRY_BACKOFFS):
-    """Generic bounded-retry wrapper shared by all four streaming complete()
-    paths. `attempt_fn()` performs exactly one full attempt -- open the
-    stream, consume it to completion, assemble and return the OpenAI-shaped
-    result dict -- and this wrapper retries the whole attempt on any
-    transient failure, following the same fixed backoff schedule the
-    non-streaming paths use.
-
-    StreamNotSupportedError is deliberately not caught here: it means the
-    gateway returned HTTP 200 with a non-SSE body (e.g. plain JSON), which no
-    amount of retrying fixes -- it must propagate so the caller can fall back
-    via its escape-hatch config flag (limits.anthropic_stream etc.) instead
-    of being silently swallowed as "just another transient error".
-
-    HTTPError is classified by status like `_http_json_post_with_retry` /
-    `_anthropic_post_with_retry` already do: 429/408/5xx are transient and
-    retried; any other 4xx fails immediately (a malformed request resends
-    identically)."""
-    attempt = 0
-    while True:
-        attempt += 1
-        try:
-            return attempt_fn()
-        except urllib.error.HTTPError as e:
-            status = e.code
-            transient = status in (429, 408) or 500 <= status < 600
-            body = getattr(e, "stream_error_body", None) or _read_http_error_body(e)
-            if not transient:
-                raise RuntimeError(f"HTTP {status} opening stream: {body}") from e
-            if attempt - 1 >= len(backoffs):
-                raise RuntimeError(f"HTTP {status} opening stream after {attempt} attempts: {body}") from e
-            time.sleep(backoffs[attempt - 1])
-        except _STREAM_RETRYABLE_EXCEPTIONS as e:
-            if attempt - 1 >= len(backoffs):
-                raise RuntimeError(f"stream failed after {attempt} attempts: {e}") from e
-            time.sleep(backoffs[attempt - 1])
+from harvest_clients.anthropic import (
+    _anthropic_post_with_retry,
+    _oai_tools_to_anthropic,
+    _assistant_to_anthropic_content,
+    _oai_messages_to_anthropic,
+    _append_user_blocks,
+    _anthropic_resp_to_oai,
+    _CACHE_CONTROL,
+    _mark_block_list_tail,
+    _mark_message_tail,
+    _inject_cache_control,
+)
+from harvest_clients.responses import (
+    _oai_tools_to_responses,
+    _oai_messages_to_responses,
+    _responses_resp_to_oai,
+)
+from harvest_clients.gemini import (
+    _gemini_generate_content_url,
+    _gemini_stream_generate_content_url,
+    _oai_tools_to_gemini,
+    _append_gemini_user_parts,
+    _gemini_tool_response_payload,
+    _oai_messages_to_gemini,
+    _gemini_resp_to_oai,
+)
 
 
 # ---------------------------------------------------------------------------
@@ -1510,1304 +1259,6 @@ def build_backends(config):
     fetch_backends = [(b, limiter_for(f"fetch:{i}", fetch_interval))
                        for i, b in enumerate(config.get("fetch_backends", []))]
     return search_backends, fetch_backends
-
-
-# ---------------------------------------------------------------------------
-# Chat client (OpenAI-compatible gateway)
-# ---------------------------------------------------------------------------
-
-class GatewayClient:
-    def __init__(self, base_url, api_key, model_id, timeout_s,
-                 stream_enabled=False, ttft_timeout=None, inter_token_timeout=None,
-                 stream_options=True):
-        self.base_url = base_url.rstrip("/")
-        self.api_key = api_key
-        self.model_id = model_id
-        self.timeout_s = timeout_s
-        # Streaming is opt-in (default False) so every existing call site --
-        # tests included -- keeps the exact non-streaming behavior it had
-        # before this parameter existed. make_client_factory is the one place
-        # that turns it on from config (limits.gateway_stream).
-        self.stream_enabled = stream_enabled
-        self.ttft_timeout = ttft_timeout
-        self.inter_token_timeout = inter_token_timeout
-        # Some OpenAI-compatible gateways only emit a final usage-bearing
-        # chunk (choices: []) if stream_options.include_usage is requested;
-        # others reject an unrecognized field outright, hence the escape
-        # hatch (limits.gateway_stream_options) rather than always-on.
-        self.stream_options = stream_options
-
-    def complete(self, messages, tools):
-        if not self.stream_enabled:
-            url = self.base_url + "/chat/completions"
-            payload = {"model": self.model_id, "messages": messages}
-            if tools:
-                payload["tools"] = tools
-            headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
-            return _http_json_post_with_retry(url, payload, headers, self.timeout_s,
-                                               transient_backoffs=_COMPLETION_RETRY_BACKOFFS)
-        return self._complete_streaming(messages, tools)
-
-    def _complete_streaming(self, messages, tools):
-        url = self.base_url + "/chat/completions"
-        payload = {"model": self.model_id, "messages": messages, "stream": True}
-        if tools:
-            payload["tools"] = tools
-        if self.stream_options:
-            payload["stream_options"] = {"include_usage": True}
-        headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
-
-        def attempt():
-            with contextlib.closing(_http_stream_post(url, payload, headers, self.ttft_timeout)) as stream:
-                resp = next(stream)
-                sock = _get_raw_socket(resp)
-                text_parts = []
-                tool_calls = {}  # index -> {"id","type","function":{"name","arguments"}}
-                finish_reason = None
-                usage = {}
-                seen_done = False
-                tightened = False
-                for _event_type, data_str in stream:
-                    if data_str is None or not data_str.strip():
-                        continue
-                    if data_str == "[DONE]":
-                        seen_done = True
-                        break
-                    chunk = json.loads(data_str)  # JSONDecodeError -> outer retry loop
-                    if chunk.get("error"):
-                        raise StreamConnectionLost(f"error frame mid-stream: {chunk['error']}")
-                    if chunk.get("usage"):
-                        usage = chunk["usage"]
-                    choices = chunk.get("choices") or []
-                    if not choices:
-                        # A choices:[] chunk carries only usage (the
-                        # include_usage final chunk) -- nothing else to do.
-                        continue
-                    choice = choices[0]
-                    delta = choice.get("delta") or {}
-                    if choice.get("finish_reason"):
-                        finish_reason = choice["finish_reason"]
-                    has_content = bool(delta.get("content")) or bool(delta.get("tool_calls"))
-                    if has_content and not tightened and sock is not None:
-                        sock.settimeout(self.inter_token_timeout)
-                        tightened = True
-                    if delta.get("content"):
-                        text_parts.append(delta["content"])
-                    for tc in (delta.get("tool_calls") or []):
-                        idx = tc.get("index", 0)
-                        slot = tool_calls.setdefault(idx, {
-                            "id": "", "type": "function",
-                            "function": {"name": "", "arguments": ""},
-                        })
-                        if tc.get("id"):
-                            slot["id"] = tc["id"]
-                        fn = tc.get("function") or {}
-                        if fn.get("name"):
-                            slot["function"]["name"] = fn["name"]
-                        if fn.get("arguments"):
-                            slot["function"]["arguments"] += fn["arguments"]
-
-            if not seen_done and finish_reason is None:
-                raise StreamConnectionLost("stream ended without [DONE] or a finish_reason")
-
-            message = {"role": "assistant"}
-            if tool_calls:
-                message["content"] = "".join(text_parts) if text_parts else None
-                message["tool_calls"] = [tool_calls[i] for i in sorted(tool_calls)]
-            else:
-                message["content"] = "".join(text_parts)
-            return {
-                "choices": [{"message": message, "finish_reason": finish_reason}],
-                "usage": usage,
-            }
-
-        return _run_stream_attempt_with_retry(attempt)
-
-
-# ---------------------------------------------------------------------------
-# Anthropic-native client (claude models only)
-#
-# claude models get a dedicated path to the gateway's Anthropic-native
-# /messages endpoint (not the OpenAI-compatible /chat/completions used for
-# gemini/gpt) so that prompt caching (cache_control) actually takes effect --
-# the OpenAI-compat conversion in the gateway silently drops cache_control,
-# so claude via that path never caches. This client speaks the exact same
-# complete(messages, tools) -> OpenAI-shaped-dict interface as GatewayClient,
-# so run_worker / judge_clusters stay identical; all protocol translation is
-# confined to the pure functions below.
-# ---------------------------------------------------------------------------
-
-def _read_http_error_body(e):
-    """Best-effort read of an HTTPError's response body so Anthropic's own
-    error diagnostics (e.g. {"error":{"message":"messages: roles must
-    alternate ..."}}) survive into the raised exception instead of being
-    swallowed as a bare status code."""
-    try:
-        return e.read().decode("utf-8", "replace")[:800]
-    except Exception:
-        return "<error body unavailable>"
-
-
-def _anthropic_post_with_retry(url, payload, headers, timeout, transient_backoffs=None):
-    """Retry wrapper for the Anthropic /messages endpoint. Unlike the OpenAI
-    gateway path, a non-transient 4xx (400/401/403/404) is NOT retried -- a
-    malformed request body fails identically on a resend, so we surface the
-    Anthropic error body immediately for debugging. Only 429/408/5xx and
-    network-layer failures follow the transient backoff schedule. HTTPError
-    is caught before URLError (its parent) so the status-bearing case isn't
-    swallowed."""
-    backoffs = transient_backoffs if transient_backoffs is not None else _GATEWAY_RETRY_BACKOFFS_TRANSIENT
-    attempt = 0
-    while True:
-        attempt += 1
-        try:
-            return _http_json_post(url, payload, headers, timeout)
-        except urllib.error.HTTPError as e:
-            status = e.code
-            transient = status in (429, 408) or 500 <= status < 600
-            body = _read_http_error_body(e)
-            if not transient:
-                raise RuntimeError(f"Anthropic HTTP {status}: {body}") from e
-            if attempt - 1 >= len(backoffs):
-                raise RuntimeError(f"Anthropic HTTP {status} after {attempt} attempts: {body}") from e
-            time.sleep(backoffs[attempt - 1])
-        except (urllib.error.URLError, socket.timeout, TimeoutError, json.JSONDecodeError) as e:
-            if attempt - 1 >= len(backoffs):
-                raise RuntimeError(f"Anthropic network error after {attempt} attempts: {e}") from e
-            time.sleep(backoffs[attempt - 1])
-
-
-def _oai_tools_to_anthropic(tools):
-    """OpenAI function-tool schema -> Anthropic tool schema. Returns None when
-    there are no tools (judge path passes tools=None) so the caller can omit
-    the field entirely rather than send tools:null / tools:[]."""
-    if not tools:
-        return None
-    out = []
-    for t in tools:
-        fn = t.get("function", {})
-        out.append({
-            "name": fn.get("name", ""),
-            "description": fn.get("description", ""),
-            "input_schema": fn.get("parameters", {"type": "object", "properties": {}}),
-        })
-    return out
-
-
-def _assistant_to_anthropic_content(msg):
-    """Convert one OpenAI assistant message into an Anthropic assistant
-    content-block list: leading text block (only if non-empty -- Anthropic
-    rejects empty text blocks) followed by one tool_use block per tool_call.
-    A plain-text assistant turn (no tool_calls) returns its string content
-    directly. tool_call arguments are JSON strings (json.dumps output on the
-    outbound side); json.loads restores the dict for tool_use.input, tolerant
-    of a malformed/empty string."""
-    tool_calls = msg.get("tool_calls") or []
-    text = msg.get("content")
-    if not tool_calls:
-        # Plain assistant answer. Anthropic needs a non-empty content; fall
-        # back to a single space if the model somehow returned empty (should
-        # not happen on a no-tool turn, but never emit an empty text block).
-        return text if isinstance(text, str) and text else " "
-    blocks = []
-    if isinstance(text, str) and text:
-        blocks.append({"type": "text", "text": text})
-    for tc in tool_calls:
-        fn = tc.get("function", {})
-        try:
-            args = json.loads(fn.get("arguments") or "{}")
-        except json.JSONDecodeError:
-            args = {}
-        blocks.append({
-            "type": "tool_use",
-            "id": tc.get("id", ""),
-            "name": fn.get("name", ""),
-            "input": args,
-        })
-    return blocks
-
-
-def _oai_messages_to_anthropic(messages):
-    """Split an OpenAI-style messages list into (system, anthropic_messages).
-
-    - role=system  -> hoisted into the top-level `system` (concatenated if
-      more than one; there is only ever one here in practice).
-    - role=user    -> passed through (content is str or a content-block list).
-    - role=assistant -> text + tool_use blocks (see helper).
-    - role=tool    -> a tool_result block; *consecutive* tool messages are
-      merged into a single user turn (Anthropic requires all tool_result
-      blocks answering the previous assistant's tool_use to live in one user
-      message). run_worker appends one role=tool per tool_call, so this run of
-      tool messages must collapse into one user turn.
-
-    Adjacent same-role user turns are also merged, mirroring the gateway's
-    "no two consecutive user messages" constraint that append_or_merge_user
-    already relies on."""
-    system_parts = []
-    anth = []
-    i = 0
-    n = len(messages)
-    while i < n:
-        msg = messages[i]
-        role = msg.get("role")
-        if role == "system":
-            content = msg.get("content")
-            if isinstance(content, str):
-                system_parts.append(content)
-            elif isinstance(content, list):
-                for blk in content:
-                    if isinstance(blk, dict) and blk.get("type") == "text":
-                        system_parts.append(blk.get("text", ""))
-            i += 1
-            continue
-        if role == "tool":
-            # Collapse the maximal run of consecutive tool messages into one
-            # user turn of tool_result blocks.
-            results = []
-            while i < n and messages[i].get("role") == "tool":
-                tmsg = messages[i]
-                results.append({
-                    "type": "tool_result",
-                    "tool_use_id": tmsg.get("tool_call_id", ""),
-                    "content": tmsg.get("content", ""),
-                })
-                i += 1
-            _append_user_blocks(anth, results)
-            continue
-        if role == "assistant":
-            anth.append({"role": "assistant", "content": _assistant_to_anthropic_content(msg)})
-            i += 1
-            continue
-        # role == user (or unknown -> treat as user)
-        _append_user_blocks(anth, msg.get("content"))
-        i += 1
-    system = "\n\n".join(p for p in system_parts if p) if system_parts else None
-    return system, anth
-
-
-def _append_user_blocks(anth, content):
-    """Append `content` as a user turn, merging into a trailing user turn if
-    present (Anthropic rejects two consecutive user messages). `content` may
-    be a str, a single block dict, or a list of blocks; the merged turn is
-    normalized to a block list when mixing shapes."""
-    if isinstance(content, list):
-        incoming = content
-    elif isinstance(content, dict):
-        incoming = [content]
-    elif content:  # non-empty string
-        incoming = [{"type": "text", "text": content}]
-    else:  # None or "" -> contribute nothing; never fabricate an empty text
-        incoming = []  # block (Anthropic rejects empty text blocks -> 400)
-    if not incoming:
-        return
-    if anth and anth[-1]["role"] == "user":
-        existing = anth[-1]["content"]
-        if isinstance(existing, str):
-            existing = [{"type": "text", "text": existing}] if existing else []
-        elif isinstance(existing, dict):
-            existing = [existing]
-        existing.extend(incoming)
-        anth[-1]["content"] = existing
-    else:
-        anth.append({"role": "user", "content": incoming})
-
-
-def _anthropic_resp_to_oai(resp):
-    """Anthropic /messages response -> OpenAI chat-completions shape so
-    _extract_message (resp["choices"][0]["message"]) and the run_worker /
-    judge_clusters loops read it unchanged.
-
-    - text blocks are concatenated into `content`.
-    - tool_use blocks become OpenAI tool_calls (arguments re-serialized to a
-      JSON string via json.dumps, matching what the OpenAI path produces).
-    - content is a string for a text-only reply (judge / final synthesis rely
-      on a str here); it is None only when the turn is purely tool_use, which
-      run_worker's `message.get("content") or ""` and the inbound converter
-      both tolerate.
-    - stop_reason is surfaced as finish_reason for observability (e.g.
-      max_tokens truncation shows up instead of silently yielding half a
-      JSON blob)."""
-    content_blocks = resp.get("content") or []
-    text_parts = []
-    tool_calls = []
-    for blk in content_blocks:
-        btype = blk.get("type")
-        if btype == "text":
-            text_parts.append(blk.get("text", ""))
-        elif btype == "tool_use":
-            tool_calls.append({
-                "id": blk.get("id", ""),
-                "type": "function",
-                "function": {
-                    "name": blk.get("name", ""),
-                    "arguments": json.dumps(blk.get("input", {}), ensure_ascii=False),
-                },
-            })
-    message = {"role": "assistant"}
-    if tool_calls:
-        message["content"] = "".join(text_parts) if text_parts else None
-        message["tool_calls"] = tool_calls
-    else:
-        message["content"] = "".join(text_parts)
-    return {
-        "choices": [{"message": message, "finish_reason": resp.get("stop_reason")}],
-        "usage": resp.get("usage", {}),
-    }
-
-
-class AnthropicGatewayClient:
-    def __init__(self, base_url, api_key, model_id, timeout_s, max_tokens, effort=None,
-                 stream_enabled=False, ttft_timeout=None, inter_token_timeout=None):
-        self.base_url = base_url.rstrip("/")
-        self.api_key = api_key
-        self.model_id = model_id
-        self.timeout_s = timeout_s
-        self.max_tokens = max_tokens
-        # effort=None preserves the exact wire behavior from before this
-        # field existed (no output_config sent at all, provider default
-        # applies -- currently "high"). make_client_factory is the one place
-        # that supplies a non-None value (config-driven, default "medium");
-        # the class itself stays decoupled from that policy choice so it can
-        # be unit-tested with either shape. See ADR-011 (research repo) --
-        # this is a deliberate global default lowering, not "old behavior
-        # preserved for old configs".
-        self.effort = effort
-        # Streaming opt-in, same contract as GatewayClient's stream_enabled --
-        # default False keeps every existing (positional-args-only) call site
-        # on the old non-streaming behavior untouched.
-        self.stream_enabled = stream_enabled
-        self.ttft_timeout = ttft_timeout
-        self.inter_token_timeout = inter_token_timeout
-
-    def _build_payload(self, messages, tools):
-        system, anth_messages = _oai_messages_to_anthropic(messages)
-        anth_tools = _oai_tools_to_anthropic(tools)
-        system = _inject_cache_control(system, anth_tools, anth_messages)
-        payload = {
-            "model": self.model_id,
-            "max_tokens": self.max_tokens,
-            "messages": anth_messages,
-        }
-        if system is not None:
-            payload["system"] = system
-        if anth_tools is not None:
-            payload["tools"] = anth_tools
-        if self.effort is not None:
-            payload["output_config"] = {"effort": self.effort}
-        return payload
-
-    def complete(self, messages, tools):
-        if not self.stream_enabled:
-            return self._complete_blocking(messages, tools)
-        return self._complete_streaming(messages, tools)
-
-    def _complete_blocking(self, messages, tools):
-        payload = self._build_payload(messages, tools)
-        url = self.base_url + "/messages"
-        headers = {
-            "Authorization": f"Bearer {self.api_key}",
-            "Content-Type": "application/json",
-            "anthropic-version": "2023-06-01",
-        }
-        resp = _anthropic_post_with_retry(url, payload, headers, self.timeout_s,
-                                          transient_backoffs=_COMPLETION_RETRY_BACKOFFS)
-        return _anthropic_resp_to_oai(resp)
-
-    def _complete_streaming(self, messages, tools):
-        payload = self._build_payload(messages, tools)
-        payload["stream"] = True
-        url = self.base_url + "/messages"
-        headers = {
-            "Authorization": f"Bearer {self.api_key}",
-            "Content-Type": "application/json",
-            "anthropic-version": "2023-06-01",
-        }
-
-        def attempt():
-            with contextlib.closing(_http_stream_post(url, payload, headers, self.ttft_timeout)) as stream:
-                resp = next(stream)
-                sock = _get_raw_socket(resp)
-                text_parts = []
-                tool_calls = []  # list, in content_block index order
-                current_tool_input_json = {}  # block_index -> accumulated partial_json
-                finish_reason = None
-                usage = {}
-                seen_message_stop = False
-                tightened = False
-                for event_type, data_str in stream:
-                    if data_str is None or not data_str.strip():
-                        continue
-                    chunk = json.loads(data_str)  # JSONDecodeError -> outer retry loop
-                    if event_type == "error":
-                        raise StreamConnectionLost(f"error event mid-stream: {chunk}")
-                    if event_type == "message_start":
-                        msg_usage = (chunk.get("message") or {}).get("usage") or {}
-                        if msg_usage:
-                            usage["input_tokens"] = msg_usage.get("input_tokens")
-                    elif event_type == "content_block_start":
-                        block = chunk.get("content_block") or {}
-                        if block.get("type") == "tool_use":
-                            idx = chunk.get("index", len(tool_calls))
-                            tool_calls.append((idx, {
-                                "id": block.get("id", ""),
-                                "type": "function",
-                                "function": {"name": block.get("name", ""), "arguments": ""},
-                            }))
-                            current_tool_input_json[idx] = ""
-                    elif event_type == "content_block_delta":
-                        delta = chunk.get("delta") or {}
-                        dtype = delta.get("type")
-                        is_content_delta = dtype in ("text_delta", "thinking_delta", "input_json_delta")
-                        if is_content_delta and not tightened and sock is not None:
-                            sock.settimeout(self.inter_token_timeout)
-                            tightened = True
-                        if dtype == "text_delta":
-                            text_parts.append(delta.get("text", ""))
-                        elif dtype == "input_json_delta":
-                            idx = chunk.get("index")
-                            if idx in current_tool_input_json:
-                                current_tool_input_json[idx] += delta.get("partial_json", "")
-                    elif event_type == "message_delta":
-                        delta_usage = chunk.get("usage") or {}
-                        if delta_usage:
-                            usage["output_tokens"] = delta_usage.get("output_tokens")
-                        raw_stop_reason = (chunk.get("delta") or {}).get("stop_reason")
-                        if raw_stop_reason:
-                            finish_reason = raw_stop_reason
-                    elif event_type == "message_stop":
-                        seen_message_stop = True
-
-            if not seen_message_stop:
-                raise StreamConnectionLost("stream ended without a message_stop event")
-
-            # Fold each tool call's accumulated partial_json into its
-            # function.arguments, matching _anthropic_resp_to_oai's
-            # json.dumps(tool_use.input)-as-a-string contract.
-            ordered_tool_calls = []
-            for idx, tc in tool_calls:
-                raw_json = current_tool_input_json.get(idx, "")
-                try:
-                    parsed_input = json.loads(raw_json) if raw_json else {}
-                except json.JSONDecodeError:
-                    parsed_input = {}
-                tc["function"]["arguments"] = json.dumps(parsed_input, ensure_ascii=False)
-                ordered_tool_calls.append(tc)
-
-            message = {"role": "assistant"}
-            if ordered_tool_calls:
-                message["content"] = "".join(text_parts) if text_parts else None
-                message["tool_calls"] = ordered_tool_calls
-            else:
-                message["content"] = "".join(text_parts)
-            return {
-                "choices": [{"message": message, "finish_reason": finish_reason}],
-                "usage": usage,
-            }
-
-        return _run_stream_attempt_with_retry(attempt)
-
-
-_CACHE_CONTROL = {"type": "ephemeral"}
-# Anthropic allows at most 4 cache breakpoints. We use 3: tools tail, system
-# tail, and the last message content block (the moving conversation
-# breakpoint). All use the default 5-minute TTL rather than the "1h" option:
-# a worker's agentic loop and the judge call all complete well inside 5
-# minutes, so the longer TTL buys nothing here while its 2x write cost and
-# the unsettled beta-header requirement for "ttl":"1h" (the docs don't state
-# whether it needs an anthropic-beta header) would only add risk. A
-# breakpoint caches the ENTIRE prefix up to and including its block, in the
-# fixed order tools -> system -> messages, so these three points let every
-# turn read the whole static prefix (tools+system) plus all prior history
-# from cache and pay full price only for the newest increment.
-
-
-def _mark_block_list_tail(blocks):
-    """Attach cache_control to the last cacheable (dict) block in a list.
-    Anthropic rejects cache_control on an empty text block, so skip a tail
-    block that is an empty-text block. No-op if there is no dict block."""
-    for blk in reversed(blocks):
-        if not isinstance(blk, dict):
-            continue
-        if blk.get("type") == "text" and not blk.get("text"):
-            continue
-        blk["cache_control"] = dict(_CACHE_CONTROL)
-        return True
-    return False
-
-
-def _mark_message_tail(messages):
-    """Move the conversation breakpoint onto the last content block of the
-    last message. String content is promoted to a one-text-block list so the
-    marker has somewhere to live (Anthropic accepts either shape); an empty
-    string is left untouched (nothing cacheable, and an empty text block is
-    rejected)."""
-    if not messages:
-        return
-    last = messages[-1]
-    content = last.get("content")
-    if isinstance(content, str):
-        if not content:
-            return
-        last["content"] = [{"type": "text", "text": content, "cache_control": dict(_CACHE_CONTROL)}]
-    elif isinstance(content, list) and content:
-        _mark_block_list_tail(content)
-
-
-def _inject_cache_control(system, tools, messages):
-    """Attach cache_control breakpoints in place for maximal prefix caching.
-
-    cache_control cannot ride on a bare string, so a non-empty system string
-    is promoted here into a single cache-marked text block list (Anthropic
-    accepts system as either a string or a block list); an already-block-list
-    system just gets its tail marked. Because this rewrites system, the value
-    is RETURNED and AnthropicGatewayClient.complete must send the returned
-    system in the payload -- tools and messages are mutated in place. An empty
-    or None system is returned untouched (nothing cacheable).
-
-    Breakpoints: tools tail + system tail (static prefix) + last message
-    content block (see _mark_message_tail). Returns the (possibly rewritten)
-    system value."""
-    if tools:
-        _mark_block_list_tail(tools)
-    if isinstance(system, str) and system:
-        system = [{"type": "text", "text": system, "cache_control": dict(_CACHE_CONTROL)}]
-    elif isinstance(system, list) and system:
-        _mark_block_list_tail(system)
-    _mark_message_tail(messages)
-    # Known limit: only the single moving message breakpoint is used, so if
-    # one turn ever appended >=20 content blocks (Anthropic's incremental-
-    # cache lookback window) the prior breakpoint would fall out of window and
-    # that one turn would miss cache. In this pipeline a turn adds ~2 blocks
-    # (one tool_use + one tool_result), so the window is never approached; a
-    # 4th "anchor" breakpoint would only pay off with heavy parallel tool use
-    # and is deliberately omitted to keep the strategy simple.
-    return system
-
-
-# ---------------------------------------------------------------------------
-# Responses-API client (gpt models only)
-#
-# gpt models get a dedicated path to the gateway's OpenAI Responses endpoint
-# (/responses) instead of /chat/completions. This is a KV-cache optimization,
-# not a capability upgrade: on this gateway (aapi.tbps.one, a new-api/one-api
-# aggregator that routes multiple upstream instances), repeated calls against
-# /chat/completions draw a random instance each time and essentially never
-# reuse a warm prefix cache, while /responses calls -- even called completely
-# statelessly (no previous_response_id, full messages resent every step,
-# exactly matching run_worker's calling pattern) -- land on the same instance
-# often enough to hit a growing prefix cache (see ADR-011 in the research
-# repo for the probe data). This is an *instance-affinity accident of this
-# particular gateway*, not a documented OpenAI Responses feature -- a
-# different gateway/provider must be re-probed before relying on it again.
-# Same complete(messages, tools) -> OpenAI-shaped-dict interface as
-# GatewayClient/AnthropicGatewayClient, so run_worker/judge_clusters/
-# run_judge_with_fallback stay unchanged; all protocol translation is
-# confined to the pure functions below.
-# ---------------------------------------------------------------------------
-
-def _oai_tools_to_responses(tools):
-    """OpenAI function-tool schema -> Responses API tool schema. Responses
-    tools are flat ({"type":"function","name":...,"parameters":...}) unlike
-    chat/completions' nested {"function": {...}} shape. Returns None for no
-    tools (judge path passes tools=None) so the caller omits the field
-    entirely, mirroring _oai_tools_to_anthropic."""
-    if not tools:
-        return None
-    out = []
-    for t in tools:
-        fn = t.get("function", {})
-        out.append({
-            "type": "function",
-            "name": fn.get("name", ""),
-            "description": fn.get("description", ""),
-            "parameters": fn.get("parameters", {"type": "object", "properties": {}}),
-        })
-    return out
-
-
-def _oai_messages_to_responses(messages):
-    """OpenAI chat-messages -> (instructions, input_items) for the Responses
-    API. Mirrors _oai_messages_to_anthropic's role-by-role walk but targets
-    Responses' flat item list instead of Anthropic's nested content blocks.
-
-    - role=system   -> hoisted into `instructions` (concatenated; there is
-      only ever one here in practice).
-    - role=user     -> a message item with an input_text content block.
-    - role=assistant -> a message item carrying any text (output_text block)
-      followed by one function_call item per tool call. function_call
-      `arguments` is passed through as-is: run_worker's outbound tool_calls
-      already carry a JSON *string* (the same value the OpenAI chat-
-      completions wire format uses), which is exactly what a Responses
-      function_call item expects -- no json.loads/json.dumps round-trip
-      needed here (unlike the Anthropic path, whose tool_use.input is a
-      dict).
-    - role=tool     -> a function_call_output item keyed by call_id
-      (run_worker's tool_call_id becomes Responses' call_id).
-
-    Unlike _oai_messages_to_anthropic there is no consecutive-turn merging
-    requirement -- Responses' `input` is a flat item list with no
-    alternating-role constraint, so tool results and user nudges are each
-    emitted as their own item without a merge step."""
-    instructions_parts = []
-    items = []
-    for msg in messages:
-        role = msg.get("role")
-        if role == "system":
-            content = msg.get("content")
-            if isinstance(content, str) and content:
-                instructions_parts.append(content)
-            continue
-        if role == "tool":
-            items.append({
-                "type": "function_call_output",
-                "call_id": msg.get("tool_call_id", ""),
-                "output": msg.get("content", ""),
-            })
-            continue
-        if role == "assistant":
-            text = msg.get("content")
-            if isinstance(text, str) and text:
-                items.append({"role": "assistant", "content": [{"type": "output_text", "text": text}]})
-            for tc in (msg.get("tool_calls") or []):
-                fn = tc.get("function", {})
-                items.append({
-                    "type": "function_call",
-                    "call_id": tc.get("id", ""),
-                    "name": fn.get("name", ""),
-                    "arguments": fn.get("arguments") or "{}",
-                })
-            continue
-        # role == user (or unknown -> treat as user)
-        content = msg.get("content")
-        text = content if isinstance(content, str) else ""
-        if text:
-            items.append({"role": "user", "content": [{"type": "input_text", "text": text}]})
-    instructions = "\n\n".join(instructions_parts) if instructions_parts else None
-    return instructions, items
-
-
-def _responses_resp_to_oai(resp):
-    """Responses API response -> OpenAI chat-completions shape so
-    _extract_message and the run_worker/judge_clusters loops read it
-    unchanged (same contract as _anthropic_resp_to_oai).
-
-    - `output` is a list of items; "message" items carry output_text content
-      blocks (concatenated into `content`), "function_call" items become
-      OpenAI tool_calls. `arguments` on a Responses function_call item is
-      already a JSON string (same wire shape OpenAI chat/completions uses),
-      so it is passed straight through -- no re-serialization, unlike the
-      Anthropic path where tool_use.input is a dict needing json.dumps.
-    - content is a string for a text-only reply (judge / final synthesis
-      rely on a str here); it is None only when the turn is purely
-      function_call, mirroring the Anthropic conversion's contract.
-    - `status` is surfaced as finish_reason for observability (e.g.
-      "incomplete" from a max_output_tokens truncation shows up instead of
-      silently yielding half a JSON blob)."""
-    output = resp.get("output") or []
-    text_parts = []
-    tool_calls = []
-    for item in output:
-        itype = item.get("type")
-        if itype == "message":
-            for blk in item.get("content") or []:
-                if isinstance(blk, dict) and blk.get("type") in ("output_text", "text"):
-                    text_parts.append(blk.get("text", ""))
-        elif itype == "function_call":
-            tool_calls.append({
-                "id": item.get("call_id", ""),
-                "type": "function",
-                "function": {
-                    "name": item.get("name", ""),
-                    "arguments": item.get("arguments") or "{}",
-                },
-            })
-    message = {"role": "assistant"}
-    if tool_calls:
-        message["content"] = "".join(text_parts) if text_parts else None
-        message["tool_calls"] = tool_calls
-    else:
-        message["content"] = "".join(text_parts)
-    return {
-        "choices": [{"message": message, "finish_reason": resp.get("status")}],
-        "usage": resp.get("usage", {}),
-    }
-
-
-class ResponsesGatewayClient:
-    def __init__(self, base_url, api_key, model_id, timeout_s, max_output_tokens,
-                 stream_enabled=False, ttft_timeout=None, inter_token_timeout=None,
-                 stream_options=True):
-        self.base_url = base_url.rstrip("/")
-        self.api_key = api_key
-        self.model_id = model_id
-        self.timeout_s = timeout_s
-        # Mirrors AnthropicGatewayClient's max_tokens: Responses' equivalent
-        # knob is max_output_tokens, and it is mandatory here for the same
-        # reason max_tokens is mandatory on the Anthropic path -- a long
-        # findings/judge JSON silently truncated by an unset/too-small output
-        # cap surfaces as a parse_findings_json failure downstream, not as an
-        # obvious "output was cut off" error. make_client_factory sources
-        # this from the same limits.completion_max_tokens value used by the
-        # Anthropic path (config-level single source of truth for "how big
-        # can a completion be").
-        self.max_output_tokens = max_output_tokens
-        # Streaming opt-in, same contract as GatewayClient/AnthropicGatewayClient.
-        self.stream_enabled = stream_enabled
-        self.ttft_timeout = ttft_timeout
-        self.inter_token_timeout = inter_token_timeout
-        # Defaults to always-on (unlike GatewayClient's stream_options, which
-        # defaults on but is meant to be flippable per-gateway) -- Responses
-        # needs the final usage-bearing event to populate the returned
-        # dict's "usage" key at all. Exposed as a constructor param anyway
-        # so make_client_factory can wire it to the same
-        # limits.gateway_stream_options escape hatch as GatewayClient, in
-        # case a gateway ever rejects the field outright.
-        self.stream_options = stream_options
-
-    def _build_payload(self, messages, tools):
-        instructions, input_items = _oai_messages_to_responses(messages)
-        resp_tools = _oai_tools_to_responses(tools)
-        payload = {
-            "model": self.model_id,
-            "input": input_items,
-            "max_output_tokens": self.max_output_tokens,
-        }
-        if instructions is not None:
-            payload["instructions"] = instructions
-        if resp_tools is not None:
-            payload["tools"] = resp_tools
-        return payload
-
-    def complete(self, messages, tools):
-        if not self.stream_enabled:
-            return self._complete_blocking(messages, tools)
-        return self._complete_streaming(messages, tools)
-
-    def _complete_blocking(self, messages, tools):
-        payload = self._build_payload(messages, tools)
-        url = self.base_url + "/responses"
-        headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
-        resp = _http_json_post_with_retry(url, payload, headers, self.timeout_s,
-                                           transient_backoffs=_COMPLETION_RETRY_BACKOFFS)
-        return _responses_resp_to_oai(resp)
-
-    def _complete_streaming(self, messages, tools):
-        payload = self._build_payload(messages, tools)
-        payload["stream"] = True
-        url = self.base_url + "/responses"
-        headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
-
-        def attempt():
-            with contextlib.closing(_http_stream_post(url, payload, headers, self.ttft_timeout)) as stream:
-                resp = next(stream)
-                sock = _get_raw_socket(resp)
-                text_parts = []
-                tool_calls = {}  # output_index -> {"id","type","function":{"name","arguments"}}
-                finish_reason = None
-                usage = {}
-                seen_completed = False
-                tightened = False
-                for _event_type, data_str in stream:
-                    if data_str is None or not data_str.strip():
-                        continue
-                    chunk = json.loads(data_str)  # JSONDecodeError -> outer retry loop
-                    ctype = chunk.get("type", "")
-                    if ctype == "error" or chunk.get("error"):
-                        raise StreamConnectionLost(f"error event mid-stream: {chunk}")
-                    if ctype.endswith(".delta") and not tightened and sock is not None:
-                        sock.settimeout(self.inter_token_timeout)
-                        tightened = True
-                    if ctype == "response.output_text.delta":
-                        text_parts.append(chunk.get("delta", ""))
-                    elif ctype in ("response.output_item.added", "response.output_item.done"):
-                        item = chunk.get("item") or {}
-                        if item.get("type") == "function_call":
-                            idx = chunk.get("output_index", len(tool_calls))
-                            slot = tool_calls.setdefault(idx, {
-                                "id": "", "type": "function",
-                                "function": {"name": "", "arguments": ""},
-                            })
-                            if item.get("call_id"):
-                                slot["id"] = item["call_id"]
-                            if item.get("name"):
-                                slot["function"]["name"] = item["name"]
-                            if ctype == "response.output_item.done" and item.get("arguments"):
-                                # Consolidated final arguments string, in case
-                                # any function_call_arguments.delta frames
-                                # were missed or coalesced upstream.
-                                slot["function"]["arguments"] = item["arguments"]
-                    elif ctype == "response.function_call_arguments.delta":
-                        idx = chunk.get("output_index")
-                        if idx is not None:
-                            slot = tool_calls.setdefault(idx, {
-                                "id": "", "type": "function",
-                                "function": {"name": "", "arguments": ""},
-                            })
-                            slot["function"]["arguments"] += chunk.get("delta", "")
-                    elif ctype in ("response.completed", "response.incomplete"):
-                        seen_completed = True
-                        response_obj = chunk.get("response") or {}
-                        usage = response_obj.get("usage") or {}
-                        finish_reason = response_obj.get("status")
-
-            if not seen_completed:
-                raise StreamConnectionLost("stream ended without a response.completed/incomplete event")
-
-            message = {"role": "assistant"}
-            ordered_tool_calls = [tool_calls[i] for i in sorted(tool_calls)]
-            if ordered_tool_calls:
-                message["content"] = "".join(text_parts) if text_parts else None
-                message["tool_calls"] = ordered_tool_calls
-            else:
-                message["content"] = "".join(text_parts)
-            return {
-                "choices": [{"message": message, "finish_reason": finish_reason}],
-                "usage": usage,
-            }
-
-        return _run_stream_attempt_with_retry(attempt)
-
-
-# ---------------------------------------------------------------------------
-# Gemini-native client (gemini models only)
-#
-# gemini models get a dedicated path to the gateway's Gemini-native
-# generateContent endpoint (not the OpenAI-compatible /chat/completions used
-# before) so that the gateway's implicit/automatic prefix caching -- the same
-# instance-affinity mechanism ADR-011 identified for gpt's /responses path --
-# has a chance to kick in. No explicit cachedContents are created here; this
-# only switches the wire protocol so a growing, resent-in-full message prefix
-# is *observable* to the gateway's own caching, whatever it turns out to be.
-# Same complete(messages, tools) -> OpenAI-shaped-dict interface as the other
-# two native clients, so run_worker/judge_clusters/run_judge_with_fallback
-# stay unchanged; all protocol translation is confined to the pure functions
-# below.
-# ---------------------------------------------------------------------------
-
-def _gemini_generate_content_url(base_url, model_id):
-    """Derive the Gemini-native generateContent URL from the gateway's
-    OpenAI-compat base_url. Unlike _search_gemini_grounding's urlsplit
-    (scheme://netloc only, which silently drops any path prefix between the
-    origin and '/v1'), this strips only a trailing '/v1' *suffix* so a
-    base_url like 'https://gw.corp.com/ai-proxy/v1' keeps its '/ai-proxy'
-    prefix intact."""
-    base = base_url.rstrip("/")
-    if base.endswith("/v1"):
-        base = base[: -len("/v1")]
-    return f"{base}/v1beta/models/{model_id}:generateContent"
-
-
-def _gemini_stream_generate_content_url(base_url, model_id):
-    """Streaming counterpart of _gemini_generate_content_url -- same base-url
-    normalization, but the streamGenerateContent method with alt=sse so the
-    gateway (assumed to speak the standard Gemini REST surface) emits
-    Server-Sent Events instead of a single buffered JSON array."""
-    base = base_url.rstrip("/")
-    if base.endswith("/v1"):
-        base = base[: -len("/v1")]
-    return f"{base}/v1beta/models/{model_id}:streamGenerateContent?alt=sse"
-
-
-def _oai_tools_to_gemini(tools):
-    """OpenAI function-tool schema -> Gemini tool schema (a single entry
-    carrying all function declarations). Returns None for no tools (judge
-    path passes tools=None) so the caller omits the field entirely,
-    mirroring _oai_tools_to_anthropic/_oai_tools_to_responses."""
-    if not tools:
-        return None
-    declarations = []
-    for t in tools:
-        fn = t.get("function", {})
-        declarations.append({
-            "name": fn.get("name", ""),
-            "description": fn.get("description", ""),
-            "parameters": fn.get("parameters", {"type": "object", "properties": {}}),
-        })
-    return [{"functionDeclarations": declarations}]
-
-
-def _append_gemini_user_parts(contents, parts):
-    """Append `parts` as a Gemini user turn, merging into a trailing user
-    turn if present. Gemini enforces alternating roles the same way
-    Anthropic does, so a run of consecutive tool results (each becoming a
-    functionResponse part) and any immediately-following user nudge must
-    collapse into one turn rather than producing back-to-back user turns."""
-    if not parts:
-        return
-    if contents and contents[-1]["role"] == "user":
-        contents[-1]["parts"].extend(parts)
-    else:
-        contents.append({"role": "user", "parts": parts})
-
-
-def _gemini_tool_response_payload(content):
-    """Parse an OpenAI tool message's `content` string into the dict Gemini's
-    functionResponse.response requires. content is JSON-decoded when
-    possible; a non-JSON string, or JSON that decodes to a non-dict (list/
-    bool/number), is wrapped as {"result": <value>} so the final payload is
-    always a dict."""
-    try:
-        parsed = json.loads(content)
-    except (json.JSONDecodeError, TypeError):
-        return {"result": content}
-    if not isinstance(parsed, dict):
-        return {"result": parsed}
-    return parsed
-
-
-def _oai_messages_to_gemini(messages):
-    """OpenAI chat-messages -> (system_instruction, contents) for the Gemini
-    generateContent API. Mirrors _oai_messages_to_anthropic's role-by-role
-    walk but targets Gemini's parts/role shape.
-
-    - role=system    -> hoisted into `system_instruction` (multiple system
-      messages are concatenated with '\\n\\n', never overwritten by a later
-      one). None when there is no system message.
-    - role=assistant -> role "model"; text content becomes a text part,
-      each tool_call becomes a functionCall part.
-    - role=tool      -> OpenAI's tool message carries no function name, but
-      Gemini's functionResponse requires one, so a forward pass records
-      {tool_call_id: function_name} from every assistant tool_call seen so
-      far; a tool message whose tool_call_id has no recorded name means the
-      upstream message sequence itself violates the OpenAI contract, and
-      that is raised explicitly rather than papered over with an empty name.
-    - role=user/other -> role "user", text part.
-
-    Consecutive user-role turns (tool results, and tool results immediately
-    followed by a user nudge) are merged via _append_gemini_user_parts,
-    mirroring _oai_messages_to_anthropic's tool-result collapsing."""
-    system_parts = []
-    contents = []
-    call_id_to_name = {}
-    for msg in messages:
-        role = msg.get("role")
-        if role == "system":
-            content = msg.get("content")
-            if isinstance(content, str) and content:
-                system_parts.append(content)
-            continue
-        if role == "assistant":
-            parts = []
-            text = msg.get("content")
-            if isinstance(text, str) and text:
-                parts.append({"text": text})
-            for tc in (msg.get("tool_calls") or []):
-                fn = tc.get("function", {})
-                name = fn.get("name", "")
-                call_id_to_name[tc.get("id")] = name
-                try:
-                    args = json.loads(fn.get("arguments") or "{}")
-                except json.JSONDecodeError:
-                    args = {}
-                parts.append({"functionCall": {"name": name, "args": args}})
-            if not parts:
-                parts = [{"text": ""}]
-            contents.append({"role": "model", "parts": parts})
-            continue
-        if role == "tool":
-            call_id = msg.get("tool_call_id")
-            name = call_id_to_name.get(call_id)
-            if name is None:
-                raise ValueError(
-                    f"tool message references unknown tool_call_id={call_id!r} "
-                    "(no matching assistant tool_call seen yet)"
-                )
-            payload = _gemini_tool_response_payload(msg.get("content", ""))
-            _append_gemini_user_parts(contents, [{"functionResponse": {"name": name, "response": payload}}])
-            continue
-        # role == user (or unknown -> treat as user)
-        content = msg.get("content")
-        text = content if isinstance(content, str) else ""
-        if text:
-            _append_gemini_user_parts(contents, [{"text": text}])
-    system_instruction = None
-    if system_parts:
-        system_instruction = {"parts": [{"text": "\n\n".join(system_parts)}]}
-    return system_instruction, contents
-
-
-def _gemini_resp_to_oai(resp):
-    """Gemini generateContent response -> OpenAI chat-completions shape so
-    _extract_message and the run_worker/judge_clusters loops read it
-    unchanged (same contract as _anthropic_resp_to_oai/_responses_resp_to_oai).
-
-    Two layers of safety-interception defense, both must not raise:
-    - prompt-level: an empty/missing `candidates` list (the whole prompt was
-      blocked) is checked *before* any `candidates[0]` indexing. The block
-      reason from promptFeedback is preserved in the returned usage dict
-      (diagnostic only; no consumer reads it today) rather than discarded.
-    - candidate-level: a candidate with finishReason=="SAFETY" or no
-      "content" key (HTTP 200, but nothing to read) degrades to empty
-      content/no tool_calls via .get() chains, never a bare KeyError.
-
-    finish_reason is semantically mapped, not passed through raw: any
-    parsed tool_calls force "tool_calls" (highest priority, regardless of
-    the raw finishReason); otherwise "STOP"->"stop", "MAX_TOKENS"->"length",
-    and any other raw value is lowercased and passed through as-is (never
-    forged into "stop")."""
-    candidates = resp.get("candidates")
-    if not candidates:
-        block_reason = (resp.get("promptFeedback") or {}).get("blockReason")
-        usage = dict(resp.get("usageMetadata") or {})
-        usage["prompt_block_reason"] = block_reason
-        message = {"role": "assistant", "content": "", "tool_calls": None}
-        return {"choices": [{"message": message, "finish_reason": "prompt_blocked"}], "usage": usage}
-
-    candidate = candidates[0]
-    raw_finish = candidate.get("finishReason")
-    content_parts = (candidate.get("content") or {}).get("parts") or []
-    text_parts = []
-    tool_calls = []
-    for part in content_parts:
-        if "text" in part:
-            text_parts.append(part["text"])
-        elif "functionCall" in part:
-            fc = part["functionCall"]
-            tool_calls.append({
-                "id": f"call_{uuid.uuid4().hex[:8]}",
-                "type": "function",
-                "function": {
-                    "name": fc.get("name", ""),
-                    "arguments": json.dumps(fc.get("args", {}), ensure_ascii=False),
-                },
-            })
-    message = {"role": "assistant", "content": "".join(text_parts)}
-    if tool_calls:
-        message["tool_calls"] = tool_calls
-        finish_reason = "tool_calls"
-    else:
-        message["tool_calls"] = None
-        if raw_finish == "STOP":
-            finish_reason = "stop"
-        elif raw_finish == "MAX_TOKENS":
-            finish_reason = "length"
-        elif raw_finish:
-            finish_reason = raw_finish.lower()
-        else:
-            finish_reason = None
-    return {
-        "choices": [{"message": message, "finish_reason": finish_reason}],
-        "usage": resp.get("usageMetadata", {}),
-    }
-
-
-class GeminiNativeGatewayClient:
-    def __init__(self, base_url, api_key, model_id, timeout_s, max_output_tokens,
-                 stream_enabled=False, ttft_timeout=None, inter_token_timeout=None):
-        self.base_url = base_url.rstrip("/")
-        self.api_key = api_key
-        self.model_id = model_id
-        self.timeout_s = timeout_s
-        self.max_output_tokens = max_output_tokens
-        # Streaming opt-in, same contract as the other three clients.
-        self.stream_enabled = stream_enabled
-        self.ttft_timeout = ttft_timeout
-        self.inter_token_timeout = inter_token_timeout
-
-    def _build_payload(self, messages, tools):
-        system_instruction, contents = _oai_messages_to_gemini(messages)
-        gemini_tools = _oai_tools_to_gemini(tools)
-        payload = {
-            "contents": contents,
-            "generationConfig": {"maxOutputTokens": self.max_output_tokens},
-        }
-        if system_instruction is not None:
-            payload["systemInstruction"] = system_instruction
-        if gemini_tools is not None:
-            payload["tools"] = gemini_tools
-        return payload
-
-    def complete(self, messages, tools):
-        if not self.stream_enabled:
-            return self._complete_blocking(messages, tools)
-        return self._complete_streaming(messages, tools)
-
-    def _complete_blocking(self, messages, tools):
-        payload = self._build_payload(messages, tools)
-        url = _gemini_generate_content_url(self.base_url, self.model_id)
-        headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
-        resp = _http_json_post_with_retry(url, payload, headers, self.timeout_s,
-                                           transient_backoffs=_COMPLETION_RETRY_BACKOFFS)
-        return _gemini_resp_to_oai(resp)
-
-    def _complete_streaming(self, messages, tools):
-        # No "stream": true on the payload -- Gemini switches to SSE purely
-        # via the :streamGenerateContent?alt=sse URL, unlike the other three
-        # providers which flip a payload field on the same endpoint.
-        payload = self._build_payload(messages, tools)
-        url = _gemini_stream_generate_content_url(self.base_url, self.model_id)
-        headers = {"Authorization": f"Bearer {self.api_key}", "Content-Type": "application/json"}
-
-        def attempt():
-            with contextlib.closing(_http_stream_post(url, payload, headers, self.ttft_timeout)) as stream:
-                resp = next(stream)
-                sock = _get_raw_socket(resp)
-                text_parts = []
-                tool_calls = []
-                usage = {}
-                finish_reason = None
-                prompt_block_reason = None
-                seen_logical_end = False
-                tightened = False
-                for _event_type, data_str in stream:
-                    if data_str is None or not data_str.strip():
-                        continue
-                    # Each SSE data frame is itself a complete
-                    # generateContent-shaped response object (not a delta
-                    # envelope like the other three providers) -- accumulate
-                    # its parts and keep only the latest usageMetadata, which
-                    # the API reports cumulatively.
-                    chunk = json.loads(data_str)  # JSONDecodeError -> outer retry loop
-                    candidates = chunk.get("candidates")
-                    if not candidates:
-                        prompt_block_reason = (chunk.get("promptFeedback") or {}).get("blockReason")
-                        if prompt_block_reason:
-                            raise StreamConnectionLost(f"prompt blocked mid-stream: {prompt_block_reason}")
-                        continue
-                    if chunk.get("usageMetadata"):
-                        usage = chunk["usageMetadata"]
-                    candidate = candidates[0]
-                    content_parts = (candidate.get("content") or {}).get("parts") or []
-                    if content_parts and not tightened and sock is not None:
-                        sock.settimeout(self.inter_token_timeout)
-                        tightened = True
-                    for part in content_parts:
-                        if "text" in part:
-                            text_parts.append(part["text"])
-                        elif "functionCall" in part:
-                            fc = part["functionCall"]
-                            tool_calls.append({
-                                "id": f"call_{uuid.uuid4().hex[:8]}",
-                                "type": "function",
-                                "function": {
-                                    "name": fc.get("name", ""),
-                                    "arguments": json.dumps(fc.get("args", {}), ensure_ascii=False),
-                                },
-                            })
-                    raw_finish = candidate.get("finishReason")
-                    if raw_finish:
-                        seen_logical_end = True
-                        if tool_calls:
-                            finish_reason = "tool_calls"
-                        elif raw_finish == "STOP":
-                            finish_reason = "stop"
-                        elif raw_finish == "MAX_TOKENS":
-                            finish_reason = "length"
-                        else:
-                            finish_reason = raw_finish.lower()
-
-            if not seen_logical_end:
-                raise StreamConnectionLost("stream ended without a candidate finishReason")
-
-            # Determine finish_reason after all frames consumed (tool_calls
-            # may arrive in frames after finishReason — see ADR-013 review #4).
-            if tool_calls:
-                finish_reason = "tool_calls"
-
-            message = {"role": "assistant", "content": "".join(text_parts)}
-            if tool_calls:
-                message["tool_calls"] = tool_calls
-            return {
-                "choices": [{"message": message, "finish_reason": finish_reason}],
-                "usage": usage,
-            }
-
-        return _run_stream_attempt_with_retry(attempt)
-
-
-def make_client_factory(config):
-    gw = config["gateway"]
-    api_key = os.environ.get(gw["api_key_env"], "")
-    # Chat-completion calls (panel + judge) run their own agentic reasoning
-    # and can legitimately take far longer than a single search/fetch call
-    # -- decoupled from limits.call_timeout_s (which bounds only the fetch/
-    # search backends) so raising one doesn't silently also relax the
-    # other. .get() with a default so an older config without this key
-    # doesn't KeyError; it just falls back to the same default declared in
-    # harvest.config.json.
-    timeout = config.get("limits", {}).get("completion_timeout_s", 600)
-    # Anthropic requires max_tokens; OpenAI treats it as optional so the
-    # OpenAI path never set it. Read from config with a generous default so
-    # large findings/judge outputs aren't truncated (truncation would surface
-    # as finish_reason=max_tokens and a half-parsed JSON). Same .get()-with-
-    # default back-compat pattern as completion_timeout_s. Also doubles as
-    # the Responses path's max_output_tokens (same "how big can a completion
-    # be" knob, one config value, see ResponsesGatewayClient).
-    max_tokens = config.get("limits", {}).get("completion_max_tokens", 16384)
-    # ADR-011: deliberate global default lowering, not a like-for-like
-    # migration. Before this key existed, AnthropicGatewayClient never sent
-    # output_config at all, so the provider's own default ("high") applied.
-    # Every existing config -- including ones written before this key was
-    # added -- now gets "medium" via this .get() default, because thinking-
-    # heavy "high" effort was identified as the main per-step latency driver
-    # for claude in the panel (see ADR-011's wall-clock investigation). This
-    # is NOT the usual back-compat contract ("old config behaves exactly as
-    # before"); it is an intentional behavior change that happens to reuse
-    # the same .get()-with-default mechanism. Config authors who want the
-    # old "high" back can set limits.claude_effort explicitly.
-    effort = config.get("limits", {}).get("claude_effort", "medium")
-    # Escape hatch for the /responses instance-affinity cache behavior: this
-    # is observed, gateway-specific behavior (see ResponsesGatewayClient's
-    # docstring), not a guaranteed OpenAI feature. If a future gateway/
-    # provider swap makes /responses worse than plain /chat/completions,
-    # config can flip this back to "chat_completions" without a code change
-    # -- gpt* then falls through to the same plain GatewayClient path gemini
-    # used before GeminiNativeGatewayClient existed, rather than needing a
-    # second personality inside ResponsesGatewayClient itself.
-    gpt_endpoint = config.get("limits", {}).get("gpt_endpoint", "responses")
-    # Same escape hatch pattern as gpt_endpoint, for the Gemini-native
-    # generateContent path (see GeminiNativeGatewayClient's docstring): if
-    # the gateway's implicit prefix caching turns out not to apply to this
-    # protocol either, config can flip back to "chat_completions" without a
-    # code change -- gemini* then falls through to the same plain
-    # GatewayClient path it used before this client existed.
-    gemini_endpoint = config.get("limits", {}).get("gemini_endpoint", "native")
-    # ADR-013: streaming fail-fast watchdog config. ttft (time-to-first-token)
-    # bounds the initial connection + wait-for-first-frame; once any content
-    # delta arrives each client's own settimeout() call tightens to the
-    # (much shorter) inter-token bound, so a stalled-mid-stream gateway is
-    # caught fast instead of hanging for the full ttft window every token.
-    # Same .get()-with-default back-compat pattern as every other key here --
-    # an older config without these keys gets the same defaults declared in
-    # harvest.config.json (streaming on, generous but bounded timeouts).
-    stream_ttft = config.get("limits", {}).get("stream_ttft_timeout_s", 300)
-    stream_inter = config.get("limits", {}).get("stream_inter_token_timeout_s", 90)
-    anthropic_stream = config.get("limits", {}).get("anthropic_stream", True)
-    gpt_stream = config.get("limits", {}).get("gpt_stream", True)
-    gemini_stream = config.get("limits", {}).get("gemini_stream", True)
-    gateway_stream = config.get("limits", {}).get("gateway_stream", True)
-    gateway_stream_options = config.get("limits", {}).get("gateway_stream_options", False)
-
-    def factory(model_id):
-        # claude models go through the Anthropic-native /messages path so
-        # prompt caching can take effect; gpt models go through the
-        # Responses path for the same reason (see ResponsesGatewayClient);
-        # gemini models go through the Gemini-native generateContent path for
-        # the same reason (see GeminiNativeGatewayClient); everything else
-        # (and gpt/gemini when their *_endpoint switch is flipped off) stays
-        # on the OpenAI-compatible /chat/completions gateway. All four share
-        # the same complete(messages, tools) interface, so
-        # run_judge_with_fallback can freely mix client types across its
-        # candidates.
-        if model_id.startswith("claude"):
-            return AnthropicGatewayClient(gw["base_url"], api_key, model_id, timeout, max_tokens, effort=effort,
-                                           stream_enabled=anthropic_stream, ttft_timeout=stream_ttft,
-                                           inter_token_timeout=stream_inter)
-        if model_id.startswith("gpt") and gpt_endpoint == "responses":
-            return ResponsesGatewayClient(gw["base_url"], api_key, model_id, timeout, max_tokens,
-                                           stream_enabled=gpt_stream, ttft_timeout=stream_ttft,
-                                           inter_token_timeout=stream_inter,
-                                           stream_options=gateway_stream_options)
-        if model_id.startswith("gemini") and gemini_endpoint == "native":
-            return GeminiNativeGatewayClient(gw["base_url"], api_key, model_id, timeout, max_tokens,
-                                              stream_enabled=gemini_stream, ttft_timeout=stream_ttft,
-                                              inter_token_timeout=stream_inter)
-        return GatewayClient(gw["base_url"], api_key, model_id, timeout,
-                              stream_enabled=gateway_stream, ttft_timeout=stream_ttft,
-                              inter_token_timeout=stream_inter, stream_options=gateway_stream_options)
-
-    return factory
-
 
 # ---------------------------------------------------------------------------
 # Panel worker driver
