+++ plugins/deep-research/scripts/harvest_clients/__init__.py
@@ -1,0 +1,8 @@
+"""harvest_clients - the four provider-specific chat/completion clients used
+by harvest.py, plus the factory that dispatches by model_id prefix."""
+
+from harvest_clients.gateway import GatewayClient
+from harvest_clients.anthropic import AnthropicGatewayClient
+from harvest_clients.responses import ResponsesGatewayClient
+from harvest_clients.gemini import GeminiNativeGatewayClient
+from harvest_clients.factory import make_client_factory
